Classical Over Representation Analysis with FOBI

This function performs a traditional over representation analysis based on hypergeometric test: classes are treated as sets of individual metabolites and all metabolites are treated as equally informative. This function uses Food-Biomarker Ontology knowledge as biological information.

ora(
  metaboliteList,
  metaboliteUniverse,
  subOntology = "food",
  pvalCutoff = 0.01,
  fobi = fobitools::fobi
)

Arguments

metaboliteList

A vector of FOBI metabolite identifiers that define the selected list of metabolites.

metaboliteUniverse

A vector of FOBI metabolite identifiers that define the universe of possible metabolites (all metabolites analyzed in the study).

subOntology

A character string specifying one of the two FOBI sub-ontologies: "food", or "biomarker".

pvalCutoff

A numeric value indicating a p-value cutoff for raw p-values generated by hypergeometric test.

fobi

FOBI table obtained with `parse_fobi()`. If this value is set to NULL, the last version of FOBI will be downloaded from GitHub.

Value

A tibble with ORA results.

References

G. Korotkevich, V. Sukhov, A. Sergushichev. Fast gene set enrichment analysis. bioRxiv (2019), doi:10.1101/060012

Pol Castellano-Escuder, Raúl González-Domínguez, David S Wishart, Cristina Andrés-Lacueva, Alex Sánchez-Pla, FOBI: an ontology to represent food intake data and associate it with metabolomic data, Database, Volume 2020, 2020, baaa033, https://doi.org/10.1093/databa/baaa033.

Author

Pol Castellano-Escuder

Examples

metaboliteUniverse <- c(fobitools::idmap$FOBI[1:200], fobitools::idmap$FOBI[400:450])
metaboliteList <- c(fobitools::idmap$FOBI[1:50], fobitools::idmap$FOBI[70:80])

# Food enrichment analysis
fobitools::ora(metaboliteList = metaboliteList, 
               metaboliteUniverse = metaboliteUniverse, 
               pvalCutoff = 1)
#> metaboliteUniverse size:  251
#> metaboliteList size:  61
#> # A tibble: 146 × 6
#>    className                classSize overlap     pval   padj overlapMetabolites
#>    <chr>                        <int>   <int>    <dbl>  <dbl> <list>            
#>  1 meat food product               16      11 0.000126 0.0184 <chr [11]>        
#>  2 soybean (whole)                 23      12 0.00228  0.111  <chr [12]>        
#>  3 egg food product                13       8 0.00369  0.111  <chr [8]>         
#>  4 peanut (whole, raw)             13       8 0.00369  0.111  <chr [8]>         
#>  5 dairy food product              19      10 0.00527  0.111  <chr [10]>        
#>  6 quinoa seed (dried)             11       7 0.00532  0.111  <chr [7]>         
#>  7 sunflower seed oil              11       7 0.00532  0.111  <chr [7]>         
#>  8 milk product (fermented)        12       7 0.0103   0.188  <chr [7]>         
#>  9 chocolate                       10       6 0.0152   0.239  <chr [6]>         
#> 10 Red meat                        28      12 0.0174   0.239  <chr [12]>        
#> # ℹ 136 more rows

# Chemical class enrichment analysis
fobitools::ora(metaboliteList = metaboliteList, 
               metaboliteUniverse = metaboliteUniverse, 
               subOntology = "biomarker", 
               pvalCutoff = 1)
#> metaboliteUniverse size:  251
#> metaboliteList size:  61
#> # A tibble: 104 × 6
#>    className                   classSize overlap   pval  padj overlapMetabolites
#>    <chr>                           <int>   <int>  <dbl> <dbl> <list>            
#>  1 4'-O-methylisoflavones              2       2 0.0583 0.893 <chr [2]>         
#>  2 Isoflavones                         2       2 0.0583 0.893 <chr [2]>         
#>  3 Benzaldehydes                       3       2 0.147  0.893 <chr [2]>         
#>  4 Carotenes                           3       2 0.147  0.893 <chr [2]>         
#>  5 Quinic acids and derivativ…         3       2 0.147  0.893 <chr [2]>         
#>  6 Long-chain fatty acids              6       3 0.156  0.893 <chr [3]>         
#>  7 1-hydroxy-2-unsubstituted …         1       1 0.243  0.893 <chr [1]>         
#>  8 1-hydroxy-4-unsubstituted …         1       1 0.243  0.893 <chr [1]>         
#>  9 2'-Hydroxy-dihydrochalcones         1       1 0.243  0.893 <chr [1]>         
#> 10 2(hydroxyphenyl)acetic aci…         1       1 0.243  0.893 <chr [1]>         
#> # ℹ 94 more rows