Classical Over Representation Analysis with FOBI

This function performs a traditional Over Representation Analysis by hypergeometric test: classes are treated as sets of individual metabolites and all metabolites are treated as equally informative. This function uses Food-Biomarker Ontology knowledge as biological information.

ora(
  metaboliteList,
  metaboliteUniverse,
  subOntology = "food",
  pvalCutoff = 0.01,
  adjust = "fdr"
)

Arguments

metaboliteList
A vector of FOBI metabolite identifiers that define the selected list of metabolites. Other accepted identifiers are: HMDB, KEGG, PubChemCID, InChIKey, InChICode, and ChemSpider.

metaboliteUniverse
A vector of FOBI metabolite identifiers that define the universe of possible metabolites. The identifiers should be of the same type as the `metaboliteList` identifiers.

subOntology
A string specifying one of the two FOBI sub-ontologies: "food", or "biomarker".

pvalCutoff
A numeric value indicating a p-value cutoff for p-values generated by the hypergeometric test.

adjust
A string indicating a multiple testing method to use. Options are: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", and "fdr".

Value

A tibble with ORA results.

References

Foroushani AB, Brinkman FS and Lynn DJ (2013). Pathway-GPS and SIGORA:identifying relevant pathways based on the over-representation of their gene-pair signatures. PeerJ, 1, pp. e229.

Pol Castellano-Escuder, Raúl González-Domínguez, David S Wishart, Cristina Andrés-Lacueva, Alex Sánchez-Pla, FOBI: an ontology to represent food intake data and associate it with metabolomic data, Database, Volume 2020, 2020, baaa033, https://doi.org/10.1093/databa/baaa033.

Author

Pol Castellano-Escuder

Examples

library(fobitools)

metaboliteUniverse <- c(fobitools::idmap$FOBI[1:200], fobitools::idmap$FOBI[400:450])
metaboliteList <- c(fobitools::idmap$FOBI[1:50], fobitools::idmap$FOBI[70:80])

# Food enrichment analysis
fobitools::ora(metaboliteList = metaboliteList, 
               metaboliteUniverse = metaboliteUniverse, 
               pvalCutoff = 1)

#> Time difference of 0.1997931 secs
#> Time difference of 0.3525572 secs
#> Time difference of 0.497699 secs

#> 
#> Mapped identifiers from  FOBI  to  metaboliteNames

#> metaboliteUniverse size:  251
#> metaboliteList size:  61

#> # A tibble: 28 x 6
#>    classId         className               success classSize   pvalue pvalueAdj
#>    <fct>           <fct>                     <dbl>     <int>    <dbl>     <dbl>
#>  1 FOODON:03301415 soybean (whole)              11        17 0.000300    0.0081
#>  2 FOODON:00001006 meat food product             9        14 0.0013      0.0176
#>  3 FOODON:00001274 egg food product              8        13 0.0037      0.0298
#>  4 FOBI:0193       Red meat                      7        11 0.0053      0.0298
#>  5 FOODON:00001256 dairy food product           10        19 0.0053      0.0298
#>  6 FOODON:03301449 potato (whole, raw)           2         2 0.0583      0.272 
#>  7 FOBI:0185       Lean meat                     4         8 0.101       0.404 
#>  8 FOODON:03411222 fish                          3         6 0.156       0.546 
#>  9 FOODON:03301702 grapefruit (whole, raw)       1         1 0.243       0.756 
#> 10 FOODON:00001046 seafood product               1         2 0.428       0.994 
#> # … with 18 more rows

# Chemical class enrichment analysis
fobitools::ora(metaboliteList = metaboliteList, 
               metaboliteUniverse = metaboliteUniverse, 
               subOntology = "biomarker", 
               pvalCutoff = 1, adjust = "fdr")

#> Time difference of 0.149215 secs
#> Time difference of 0.305918 secs

#> 
#> Mapped identifiers from  FOBI  to  metaboliteNames

#> metaboliteUniverse size:  251
#> metaboliteList size:  61

#> # A tibble: 49 x 6
#>    classId    className                       success classSize pvalue pvalueAdj
#>    <fct>      <fct>                             <dbl>     <int>  <dbl>     <dbl>
#>  1 FOBI:0401… Isoflavones                           2         2 0.0583     0.407
#>  2 FOBI:0401… 4'-O-methylisoflavones                2         2 0.0583     0.407
#>  3 FOBI:0401… Flavonols                             2         3 0.147      0.407
#>  4 FOBI:0401… Carotenes                             2         3 0.147      0.407
#>  5 FOBI:0402… Quinic acids and derivatives          2         3 0.147      0.407
#>  6 FOBI:0402… Benzaldehydes                         2         3 0.147      0.407
#>  7 FOBI:0401… Long-chain fatty acids                3         6 0.156      0.407
#>  8 FOBI:0400… Aniline and substituted anilin…       1         1 0.243      0.407
#>  9 FOBI:0400… Indolyl carboxylic acids and d…       1         1 0.243      0.407
#> 10 FOBI:0400… 1-hydroxy-4-unsubstituted benz…       1         1 0.243      0.407
#> # … with 39 more rows