Text Mining Pipeline to Annotate Free Nutritional Text with FOBI

This function provides a text mining pipeline to map nutritional free text to Food-Biomarker Ontology. This pipeline is composed of five sequential layers to map food items to FOBI with the maximum accuracy as possible.

annotate_foods(foods, similarity = 0.85)

Arguments

foods
A two column data frame. First column must contain the ID (should be unique) and the second column must contain food items (it can be a word or a string).

similarity
Numeric between 0 (low) and 1 (high). This value indicates the semantic similarity cutoff used at the last layer of the text mining pipeline. 1 = exact match; 0 = very poor match. Values below 0.85 are not recommended.

Value

A list containing two tibble objects: annotated and unannotated food items.

References

Pol Castellano-Escuder, Raúl González-Domínguez, David S Wishart, Cristina Andrés-Lacueva, Alex Sánchez-Pla, FOBI: an ontology to represent food intake data and associate it with metabolomic data, Database, Volume 2020, 2020, baaa033, https://doi.org/10.1093/databa/baaa033.

Author

Pol Castellano-Escuder

Examples

# Free text annotation in FOBI
free_text <- data.frame(id = c(101, 102, 103, 104),
                        text = c("Yesterday I ate eggs and bacon with a butter toast and black tea", 
                                 "Crisp bread and rice crackers with wholegrain", 
                                 "Beef and veal, one apple", "pizza without meat"))
annotate_foods(free_text)

#> 100% annotated
#> 5.726 sec elapsed

#> $annotated
#> # A tibble: 17 x 4
#>    FOOD_ID FOOD_NAME                              FOBI_ID      FOBI_NAME        
#>    <chr>   <chr>                                  <chr>        <chr>            
#>  1 101     Yesterday I ate eggs and bacon with a… FOODON:0000… bacon food produ…
#>  2 101     Yesterday I ate eggs and bacon with a… FOODON:0331… butter           
#>  3 101     Yesterday I ate eggs and bacon with a… FOODON:0000… egg food product 
#>  4 101     Yesterday I ate eggs and bacon with a… FOODON:0000… tea based bevera…
#>  5 101     Yesterday I ate eggs and bacon with a… FOODON:0331… tea food product 
#>  6 101     Yesterday I ate eggs and bacon with a… FOODON:0330… toast            
#>  7 101     Yesterday I ate eggs and bacon with a… FOBI:07511   whole toast      
#>  8 102     Crisp bread and rice crackers with wh… FOODON:0000… bread food produ…
#>  9 102     Crisp bread and rice crackers with wh… FOODON:0000… cracker          
#> 10 102     Crisp bread and rice crackers with wh… FOBI:07522   crisp            
#> 11 102     Crisp bread and rice crackers with wh… FOODON:0330… rice grain food …
#> 12 102     Crisp bread and rice crackers with wh… FOBI:07510   whole bread      
#> 13 102     Crisp bread and rice crackers with wh… FOBI:0148    wholegrain       
#> 14 103     Beef and veal, one apple               FOODON:0000… apple (whole)    
#> 15 103     Beef and veal, one apple               FOODON:0000… beef food product
#> 16 103     Beef and veal, one apple               FOODON:0000… veal             
#> 17 104     pizza without meat                     FOBI:007956  vegetarian pizza 
#> 
#> $unannotated
#> # A tibble: 0 x 2
#> # … with 2 variables: FOOD_ID <dbl>, FOOD_NAME <chr>
#>