Performs a preranked metabolite set enrichment analysis (MSEA)

This function performs a MSEA based on the adaptive multilevel splitting Monte Carlo approach.

msea(
  metaboliteRanks,
  subOntology = "food",
  pvalCutoff = 0.01,
  fobi = fobitools::fobi
)

Arguments

metaboliteRanks
A named vector of FOBI metabolite identifiers with their metabolite-level stats.

subOntology
A character string specifying one of the two FOBI sub-ontologies: "food", or "biomarker".

pvalCutoff
A numeric value indicating a p-value cutoff for raw p-values generated by MSEA.

fobi
FOBI table obtained with `parse_fobi()`. If this value is set to NULL, the last version of FOBI will be downloaded from GitHub.

Value

A tibble with MSEA results.

References

G. Korotkevich, V. Sukhov, A. Sergushichev. Fast gene set enrichment analysis. bioRxiv (2019), doi:10.1101/060012

Pol Castellano-Escuder, Raúl González-Domínguez, David S Wishart, Cristina Andrés-Lacueva, Alex Sánchez-Pla, FOBI: an ontology to represent food intake data and associate it with metabolomic data, Database, Volume 2020, 2020, baaa033, https://doi.org/10.1093/databa/baaa033.

Author

Pol Castellano-Escuder

Examples

metabolites <- c(fobitools::idmap$FOBI[1:49], fobitools::idmap$FOBI[70:80])
random_pvals <- c(runif(n = length(metabolites)*0.3, min = 0.001, max = 0.05), runif(n = length(metabolites)*0.7, min = 0.05, max = 0.99))
names(random_pvals) <- metabolites
metaboliteRanks <- random_pvals[order(random_pvals)]

# Food enrichment analysis
fobitools::msea(metaboliteRanks = metaboliteRanks, 
                pvalCutoff = 1)

#> # A tibble: 28 x 8
#>    className              classSize log2err    ES   NES   pval  padj leadingEdge
#>    <chr>                      <int>   <dbl> <dbl> <dbl>  <dbl> <dbl> <list>     
#>  1 soft drink                     1  0.381  1      2.05 0.0155 0.217 <chr [1]>  
#>  2 sugar                          1  0.381  1      2.05 0.0155 0.217 <chr [1]>  
#>  3 potato (whole, raw)            2  0.241  0.929  1.69 0.0350 0.326 <chr [2]>  
#>  4 dairy food product            10  0.106  0.601  1.21 0.158  0.819 <chr [8]>  
#>  5 Whole grain cereal pr…        26  0.102  0.546  1.18 0.166  0.819 <chr [18]> 
#>  6 citrus fruit food pro…         1  0.0818 0.746  1.53 0.238  0.819 <chr [1]>  
#>  7 tomato (whole, raw)            1  0.0818 0.746  1.53 0.238  0.819 <chr [1]>  
#>  8 grapefruit (whole, ra…         1  0.0742 0.712  1.46 0.275  0.819 <chr [1]>  
#>  9 egg food product               8  0.0631 0.554  1.09 0.344  0.819 <chr [6]>  
#> 10 nut (whole or part)           25  0.0576 0.487  1.05 0.386  0.819 <chr [16]> 
#> # … with 18 more rows

# Chemical class enrichment analysis
fobitools::msea(metaboliteRanks = metaboliteRanks, 
                subOntology = "biomarker", 
                pvalCutoff = 1)

#> # A tibble: 48 x 8
#>    className              classSize log2err    ES   NES   pval  padj leadingEdge
#>    <chr>                      <int>   <dbl> <dbl> <dbl>  <dbl> <dbl> <list>     
#>  1 Indole-3-acetic acid …         1  0.204  0.983  1.96 0.0480 0.863 <chr [1]>  
#>  2 Benzoxazinones                 1  0.174  0.966  1.93 0.0649 0.863 <chr [1]>  
#>  3 1-hydroxy-4-unsubstit…         1  0.147  0.949  1.89 0.0879 0.863 <chr [1]>  
#>  4 Solanidines and deriv…         1  0.147  0.932  1.86 0.0879 0.863 <chr [1]>  
#>  5 O-glycosyl compounds           2  0.146  0.879  1.55 0.0899 0.863 <chr [2]>  
#>  6 Tyrosine and derivati…         1  0.120  0.898  1.79 0.126  0.891 <chr [1]>  
#>  7 Acyclic diterpenoids           1  0.102  0.847  1.69 0.168  0.891 <chr [1]>  
#>  8 Isoflavones                    2  0.0975 0.789  1.39 0.180  0.891 <chr [2]>  
#>  9 Xanthines                      1  0.0920 0.831  1.66 0.198  0.891 <chr [1]>  
#> 10 Catechin gallates              1  0.0876 0.814  1.62 0.214  0.891 <chr [1]>  
#> # … with 38 more rows